# back/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "fantasy对外",
        "title": "Fantasy Bounty API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/supplier/capabilities": {
            "post": {
                "description": "更新供应商的机器设备数量（增量更新）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "supplier"
                ],
                "summary": "更新机器能力",
                "parameters": [
                    {
                        "description": "机器能力信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/supplier.UpdateCapabilitiesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/supplier.CapabilitiesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/supplier.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/supplier/profile": {
            "get": {
                "description": "获取当前用户的供应商档案信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "supplier"
                ],
                "summary": "获取供应商档案",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/supplier.SupplierProfileResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/supplier.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "创建或更新供应商的企业信息和机器能力",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "supplier"
                ],
                "summary": "创建/更新供应商档案",
                "parameters": [
                    {
                        "description": "供应商信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/supplier.CreateSupplierRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/supplier.SupplierProfileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/supplier.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "supplier.CapabilitiesResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "supplier.CreateSupplierRequest": {
            "type": "object",
            "required": [
                "companyName",
                "companyType"
            ],
            "properties": {
                "capabilities": {
                    "description": "机器能力 {a:1, b:2}",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "companyName": {
                    "type": "string"
                },
                "companyType": {
                    "type": "string"
                }
            }
        },
        "supplier.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "supplier.SupplierProfile": {
            "type": "object",
            "properties": {
                "companyName": {
                    "description": "企业名称",
                    "type": "string"
                },
                "companyType": {
                    "description": "企业类型",
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "description": "equals users.id",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "description": "关联用户，唯一索引",
                    "type": "string"
                }
            }
        },
        "supplier.SupplierProfileResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/supplier.SupplierProfile"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "supplier.UpdateCapabilitiesRequest": {
            "type": "object",
            "required": [
                "capabilities"
            ],
            "properties": {
                "capabilities": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                }
            }
        }
    }
}